+++ scripts/whitespace.js
@@ -86,6 +86,18 @@ const rules = (i, prefix) => [
   ruleset(`html[dir="rtl"] ${classname(prefix, `marginEnd${i}`)}`, {
     'margin-left': bt(i),
   }),
+  ruleset(`html:not([dir="rtl"]) ${classname(prefix, `marginStartN${i}`)}`, {
+    'margin-left': bt(-i),
+  }),
+  ruleset(`html:not([dir="rtl"]) ${classname(prefix, `marginEndN${i}`)}`, {
+    'margin-right': bt(-i),
+  }),
+  ruleset(`html[dir="rtl"] ${classname(prefix, `marginStartN${i}`)}`, {
+    'margin-right': bt(-i),
+  }),
+  ruleset(`html[dir="rtl"] ${classname(prefix, `marginEndN${i}`)}`, {
+    'margin-left': bt(-i),
+  }),
   ruleset(classname(prefix, `marginTop${i}`), {
     'margin-top': bt(i),
   }),